+++ src/main.js
@@ -15,14 +15,14 @@ const FILMS_COUNT = 20;
 const films = new Array(FILMS_COUNT).fill().map(generateFilm);
 const filters = generateFilters(films);
 
-const render = (container, template, place = 'beforeend') => {
-  container.insertAdjacentHTML(place, template);
-};
-
 const siteHeaderElement = document.querySelector('.header');
 const siteMainElement = document.querySelector('.main');
 const siteFooterElement = document.querySelector('.footer');
 const footerStatistics = siteFooterElement.querySelector('.footer__statistics');
+
+const render = (container, template, place = 'beforeend') => {
+  container.insertAdjacentHTML(place, template);
+};
 
 render(siteHeaderElement, createUserRankTemplate(filters));
 render(siteMainElement, createSiteMenuTemplate(filters), 'afterbegin');
@@ -31,45 +31,33 @@ render(footerStatistics, createMoviesCounterTemplate(films));
 render(siteFooterElement, createFilmPopupTemplate(films[15]), 'afterend');
 
 const filmListElement = siteMainElement.querySelector('.films-list__container');
+const extraFilmListsElements = siteMainElement.querySelectorAll('.films-list--extra .films-list__container');
 
-const onShowMoreButtonClick = (evt) => {
-  evt.preventDefault();
+for (let i = 0; i < Math.min(films.length, FILMS_CARD_COUNTER); i++) {
+  render(filmListElement, createFilmCardTemplate(films[i]));
+}
 
-  const filmsCount = films.length;
-  const lastFilmId = filmListElement.children.length;
-
-  if (lastFilmId + FILMS_CARD_COUNTER < filmsCount) {
-    for (let i = lastFilmId; i < lastFilmId + FILMS_CARD_COUNTER; i++) {
-      render(filmListElement, createFilmCardTemplate(films[i]));
-    }
-  } else {
-    for (let i = lastFilmId; i < filmsCount; i++) {
-      render(filmListElement, createFilmCardTemplate(films[i]));
-    }
-
-    const showMoreButton = siteMainElement.querySelector('.films-list__show-more');
-
-    showMoreButton.remove();
-  }
-};
-
-if (films.length <= FILMS_CARD_COUNTER) {
-  for (let i = 0; i < films.length; i++) {
-    render(filmListElement, createFilmCardTemplate(films[i]));
-  }
-} else {
-  for (let i = 0; i < FILMS_CARD_COUNTER; i++) {
-    render(filmListElement, createFilmCardTemplate(films[i]));
-  }
+if (films.length > FILMS_CARD_COUNTER) {
+  let renderedFilmsCount = FILMS_CARD_COUNTER;
 
   render(filmListElement, createShowMoreButtonTemplate(), 'afterend');
 
   const showMoreButton = siteMainElement.querySelector('.films-list__show-more');
 
-  showMoreButton.addEventListener('click', onShowMoreButtonClick);
-}
+  showMoreButton.addEventListener('click', (evt) => {
+    evt.preventDefault();
 
-const extraFilmListsElements = siteMainElement.querySelectorAll('.films-list--extra .films-list__container');
+    films
+      .slice(renderedFilmsCount, renderedFilmsCount + FILMS_CARD_COUNTER)
+      .forEach((film) => render(filmListElement, createFilmCardTemplate(film)));
+
+    renderedFilmsCount += FILMS_CARD_COUNTER;
+
+    if (renderedFilmsCount >= films.length) {
+      showMoreButton.remove();
+    }
+  });
+}
 
 for (let i = 0; i < extraFilmListsElements.length; i++) {
   for (let j = 0; j < EXTRA_FILMS_COUNTER; j++) {
