+++ src/mock/mock-data.js
@@ -1,0 +1,102 @@
+const posters = [
+  './images/posters/made-for-each-other.png',
+  './images/posters/popeye-meets-sinbad.png',
+  './images/posters/sagebrush-trail.jpg',
+  './images/posters/santa-claus-conquers-the-martians.jpg',
+  './images/posters/the-dance-of-life.jpg',
+  './images/posters/the-great-flamarion.jpg',
+  './images/posters/the-man-with-the-golden-arm.jpg',
+];
+
+const titles = [
+  'The Dance of Life',
+  'Sagebrush Trail',
+  'The Man with the Golden Arm',
+  'Santa Claus Conquers the Martians',
+  'Popeye the Sailor Meets Sindbad the Sailor',
+  'The Great Flamarion',
+  'Made for Each Other',
+];
+
+const descriptions = [
+  'Lorem ipsum dolor sit amet, consectetur adipiscing elit. ',
+  'Cras aliquet varius magna, non porta ligula feugiat eget. ',
+  'Fusce tristique felis at fermentum pharetra. ',
+  'Aliquam id orci ut lectus varius viverra. ',
+  'Nullam nunc ex, convallis sed finibus eget, sollicitudin eget ante. ',
+  'Phasellus eros mauris, condimentum sed nibh vitae, sodales efficitur ipsum. ',
+  'Sed blandit, eros vel aliquam faucibus, purus ex euismod diam, eu luctus nunc ante ut dui. ',
+  'Sed sed nisi sed augue convallis suscipit in sed felis. ',
+  'Aliquam erat volutpat. ',
+  'Nunc fermentum tortor ac porta dapibus. ',
+  'In rutrum ac purus sit amet tempus. ',
+];
+
+const directors = [
+  'Tom Ford',
+  'Anthony Mann',
+  'Steven Spielberg',
+  'Martin Scorsese',
+  'Alfred Hitchcock',
+];
+
+const writers = [
+  'Anne Wigton',
+  'Heinz Herald',
+  'Richard Weil',
+  'Asghar Farhadi',
+  'Eric Roth',
+  'Aaron Sorkin',
+];
+
+const actors = [
+  'Erich von Stroheim',
+  'Mary Beth Hughes',
+  'Dan Duryea',
+  'Marlon Brando',
+  'Jack Nicholson',
+  'James Stewart',
+  'Humphrey Bogart',
+];
+
+const countries = [
+  'USA',
+  'Australia',
+  'Canada',
+  'Costa Rica',
+  'France',
+];
+
+const genres = [
+  'Drama',
+  'Film-Noir',
+  'Mystery',
+  'Cartoon',
+  'Comedy',
+  'Western',
+  'Musical',
+];
+
+const texts = [
+  'Interesting setting and a good cast',
+  'Booooooooooring',
+  'Very very old. Meh',
+  'Almost two hours? Seriously?',
+  'Great movie!',
+];
+
+const emotions = [
+  'angry',
+  'puke',
+  'sleeping',
+  'smile',
+];
+
+const autors = [
+  'John Doe',
+  'Tim Macoveev',
+  'Ilya O\'Reilly',
+];
+
+
+export {posters, titles, descriptions, directors, writers, actors, countries, genres, texts, emotions, autors};
