+++ src/view/film-popup.js
@@ -4,10 +4,6 @@ import {getTimeFormat, checkList} from '../utils.js';
 const createFilmPopupTemplate = (film) => {
   const {title, alternative_title, rating, age_rating, director, writers, actors, date, runTime, country, genres, poster, description, comments, isWatchlist, isWatched, isFavorite} = film;
 
-  const getGenresCount = (genres) => {
-    return genres.length > 1 ? 'Genres' : 'Genre';
-  };
-
   const renderGenres = () => {
     return genres.map((genre) => {
       return `<span class="film-details__genre">${genre}</span>`;
@@ -83,7 +79,7 @@ const createFilmPopupTemplate = (film) => {
                 <td class="film-details__cell">${country}</td>
               </tr>
               <tr class="film-details__row">
-                <td class="film-details__term">${getGenresCount(genres)}</td>
+                <td class="film-details__term">${genres.length > 1 ? 'Genres' : 'Genre'}</td>
                 <td class="film-details__cell">${renderGenres()}</td>
               </tr>
             </table>
