+++ src/view/site-menu.js
@@ -1,14 +1,13 @@
 const createSiteMenuTemplate = (filters) => {
-  const createFiltersTemplate = () => {
-    return filters.map((filter) => {
-      return `<a href="#${filter.name}" class="main-navigation__item">${filter.name} <span class="main-navigation__item-count">${filter.count}</span></a>`;
-    }).join(' ');
-  };
+  const FiltersTemplate = filters.map((filter) => {
+    return `<a href="#${filter.name}" class="main-navigation__item">${filter.name} <span class="main-navigation__item-count">${filter.count}</span></a>`;
+  }).join(' ');
+
 
   return `<nav class="main-navigation">
     <div class="main-navigation__items">
       <a href="#all" class="main-navigation__item main-navigation__item--active">All movies</a>
-      ${createFiltersTemplate()}
+      ${FiltersTemplate}
     </div>
     <a href="#stats" class="main-navigation__additional">Stats</a>
   </nav>
